requirements.txt
dash>=2.16.0
plotly>=5.20.0
requests>=2.31.0
cachetools>=5.3.3
python-dotenv>=1.0.1
gunicorn>=21.2.0

